refactor: rename auth to accounts
