+++ src/modules/accounts/services/auth/helpers/handle-auth-provider.ts
@@ -1,8 +1,8 @@
 import { AuthProvider, UserRoles } from '@types';
 
-import { UserModel } from '../../entities';
-import { userRepository } from '../../repositories';
-import { handleAuthError, handleError } from '../../utils';
+import { UserModel } from '../../../entities';
+import { userRepository } from '../../../repositories';
+import { handleAuthError, handleError } from '../../../utils';
 
 import { createAuthPayload } from './create-auth-payload';
