+++ src/modules/accounts/services/auth/register.ts
@@ -1,8 +1,8 @@
 import { UserInput, UserRoles } from '@types';
 
-import { handleError } from '../utils';
-import { UserModel } from '../entities';
-import { userRepository } from '../repositories';
+import { handleError } from '../../utils';
+import { UserModel } from '../../entities';
+import { userRepository } from '../../repositories';
 
 import { createLoginLink } from './create-login-link';
