+++ src/modules/accounts/services/auth/verify-login-link.ts
@@ -1,5 +1,5 @@
-import { handleAuthError, handleError } from '../utils';
-import { loginLinkRepository, userRepository } from '../repositories';
+import { handleAuthError, handleError } from '../../utils';
+import { loginLinkRepository, userRepository } from '../../repositories';
 
 import { createAuthPayload } from './helpers';
