+++ src/modules/accounts/services/index.ts
@@ -1,0 +1,2 @@
+export * from './auth';
+export { updateUser } from './update-user';
