+++ src/modules/index.ts
@@ -1,2 +1,2 @@
-export * from './auth/services';
+export * from './accounts/services';
 export * from './example-feature/services/example.service';
