+++ src/shared/middlewares/auth.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
 
 import { UserRole } from '@types';
-import { handleAuthError, parseAuth } from '../../modules/auth/utils';
+import { handleAuthError, parseAuth } from '../../modules/accounts/utils';
 
 export { requireAuth };
