+++ src/shared/middlewares/index.ts
@@ -1,2 +1,2 @@
-export * from './auth.middleware';
+export * from './auth';
 export * from './morgan.middleware';
